+++ lib/eve_online/esi/models/skill_queue_entry.rb
@@ -1,15 +1,7 @@
-require 'active_support/time'
-
 module EveOnline
   module ESI
     module Models
-      class SkillQueueEntry
-        attr_reader :options
-
-        def initialize(options)
-          @options = options
-        end
-
+      class SkillQueueEntry < Base
         def as_json
           {
             skill_id: skill_id,
@@ -24,35 +16,39 @@ module EveOnline
         end
 
         def skill_id
-          options.fetch('skill_id')
+          options['skill_id']
         end
 
         def finished_level
-          options.fetch('finished_level')
+          options['finished_level']
         end
 
         def queue_position
-          options.fetch('queue_position')
+          options['queue_position']
         end
 
         def finish_date
-          ActiveSupport::TimeZone['UTC'].parse(options.fetch('finish_date'))
+          finish_date = options['finish_date']
+
+          parse_datetime_with_timezone(finish_date) if finish_date
         end
 
         def start_date
-          ActiveSupport::TimeZone['UTC'].parse(options.fetch('start_date'))
+          start_date = options['start_date']
+
+          parse_datetime_with_timezone(start_date) if start_date
         end
 
         def training_start_sp
-          options.fetch('training_start_sp')
+          options['training_start_sp']
         end
 
         def level_end_sp
-          options.fetch('level_end_sp')
+          options['level_end_sp']
         end
 
         def level_start_sp
-          options.fetch('level_start_sp')
+          options['level_start_sp']
         end
       end
     end