+++ spec/eve_online/esi/models/asset_spec.rb
@@ -1,17 +1,17 @@
 require 'spec_helper'
 
 describe EveOnline::ESI::Models::Asset do
+  let(:options) { double }
+
+  subject { described_class.new(options) }
+
+  it { should be_a(EveOnline::ESI::Models::Base) }
+
   describe '#initialize' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
     its(:options) { should eq(options) }
   end
 
   describe '#as_json' do
-    let(:options) { double }
-
     let(:assets) { described_class.new(options) }
 
     before { expect(assets).to receive(:type_id).and_return(2629) }
@@ -46,134 +46,43 @@ describe EveOnline::ESI::Models::Asset do
   end
 
   describe '#type_id' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['type_id']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('type_id')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('type_id') }
 
     specify { expect { subject.type_id }.not_to raise_error }
   end
 
   describe '#location_id' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['location_id']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('location_id')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('location_id') }
 
     specify { expect { subject.location_id }.not_to raise_error }
   end
 
   describe '#location_type' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['location_type']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('location_type')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('location_type') }
 
     specify { expect { subject.location_type }.not_to raise_error }
   end
 
   describe '#item_id' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['item_id']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('item_id')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('item_id') }
 
     specify { expect { subject.item_id }.not_to raise_error }
   end
 
   describe '#location_flag' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['location_flag']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('location_flag')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('location_flag') }
 
     specify { expect { subject.location_flag }.not_to raise_error }
   end
 
   describe '#is_singleton' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['is_singleton']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('is_singleton')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('is_singleton') }
 
     specify { expect { subject.is_singleton }.not_to raise_error }
   end
 
   describe '#quantity' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['quantity']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('quantity')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('quantity') }
 
     specify { expect { subject.quantity }.not_to raise_error }
   end