+++ spec/eve_online/esi/models/loyalty_point_spec.rb
@@ -1,17 +1,17 @@
 require 'spec_helper'
 
 describe EveOnline::ESI::Models::LoyaltyPoint do
+  let(:options) { double }
+
+  subject { described_class.new(options) }
+
+  it { should be_a(EveOnline::ESI::Models::Base) }
+
   describe '#initialize' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
     its(:options) { should eq(options) }
   end
 
   describe '#as_json' do
-    let(:options) { double }
-
     let(:loyalty_point) { described_class.new(options) }
 
     before { expect(loyalty_point).to receive(:corporation_id).and_return(1_000_035) }
@@ -26,39 +26,13 @@ describe EveOnline::ESI::Models::LoyaltyPoint do
   end
 
   describe '#corporation_id' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['corporation_id']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('corporation_id')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('corporation_id') }
 
     specify { expect { subject.corporation_id }.not_to raise_error }
   end
 
   describe '#loyalty_points' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['loyalty_points']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('loyalty_points')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('loyalty_points') }
 
     specify { expect { subject.loyalty_points }.not_to raise_error }
   end