+++ spec/eve_online/esi/models/skill_spec.rb
@@ -1,17 +1,17 @@
 require 'spec_helper'
 
 describe EveOnline::ESI::Models::Skill do
+  let(:options) { double }
+
+  subject { described_class.new(options) }
+
+  it { should be_a(EveOnline::ESI::Models::Base) }
+
   describe '#initialize' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
     its(:options) { should eq(options) }
   end
 
   describe '#as_json' do
-    let(:options) { double }
-
     let(:skill) { described_class.new(options) }
 
     before { expect(skill).to receive(:skill_id).and_return(22_536) }
@@ -30,58 +30,19 @@ describe EveOnline::ESI::Models::Skill do
   end
 
   describe '#skill_id' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['skill_id']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('skill_id')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('skill_id') }
 
     specify { expect { subject.skill_id }.not_to raise_error }
   end
 
   describe '#skillpoints_in_skill' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['skillpoints_in_skill']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('skillpoints_in_skill')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('skillpoints_in_skill') }
 
     specify { expect { subject.skillpoints_in_skill }.not_to raise_error }
   end
 
   describe '#current_skill_level' do
-    let(:options) { double }
-
-    subject { described_class.new(options) }
-
-    before do
-      #
-      # subject.options['current_skill_level']
-      #
-      expect(subject).to receive(:options) do
-        double.tap do |a|
-          expect(a).to receive(:[]).with('current_skill_level')
-        end
-      end
-    end
+    before { expect(options).to receive(:[]).with('current_skill_level') }
 
     specify { expect { subject.current_skill_level }.not_to raise_error }
   end